+++ LtiLibrary.AspNetCore.Tests/Outcomes/v2/ResultsController.cs
@@ -3,7 +3,6 @@ using LtiLibrary.AspNetCore.Extensions;
 using LtiLibrary.AspNetCore.Outcomes.v2;
 using LtiLibrary.NetCore.Common;
 using LtiLibrary.NetCore.Lis.v2;
-using LtiLibrary.NetCore.Outcomes.v2;
 using Microsoft.AspNetCore.Http;
 
 namespace LtiLibrary.AspNetCore.Tests.Outcomes.v2
