+++ LtiLibrary.AspNetCore/Outcomes/v2/LineItemModelBinder.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.IO;
 using System.Threading.Tasks;
-using LtiLibrary.NetCore.Outcomes.v2;
+using LtiLibrary.NetCore.Lis.v2;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Newtonsoft.Json;
