+++ LtiLibrary.NetCore/Lis/v2/ResultMembershipSubject.cs
@@ -1,8 +1,7 @@
 ﻿using LtiLibrary.NetCore.Common;
-using LtiLibrary.NetCore.Lis.v2;
 using Newtonsoft.Json;
 
-namespace LtiLibrary.NetCore.Outcomes.v2
+namespace LtiLibrary.NetCore.Lis.v2
 {
     /// <summary>
     /// Represents an IMS ResultMembershipSubject object.
